Degree Corrected Latent Block Model for bipartite graph class

An S4 class to represent a degree corrected stochastic block model for co_clustering of bipartite graph. Such model can be used to cluster graph vertex, and model a bipartite graph adjacency matrix \(X\) with the following generative model : $$ \pi \sim Dirichlet(\alpha)$$ $$ Z_i^r \sim \mathcal{M}(1,\pi^r)$$ $$ Z_j^c \sim \mathcal{M}(1,\pi^c)$$ $$ \theta_{kl} \sim Exponential(p)$$ $$ \gamma_i^r\sim \mathcal{U}(S_k)$$ $$ \gamma_i^c\sim \mathcal{U}(S_l)$$ $$ X_{ij}|Z_{ik}^cZ_{jl}^r=1 \sim \mathcal{P}(\gamma_i^r\theta_{kl}\gamma_j^c)$$ The individuals parameters \(\gamma_i^r,\gamma_j^c\) allow to take into account the node degree heterogeneity. These parameters have uniform priors over simplex \(S_k\). These classes mainly store the prior parameters value \(\alpha,p\) of this generative model. The DcLbm-class must be used when fitting a simple Diagonal Gaussian Mixture Model whereas the DcLbmPrior-class must be sued when fitting a CombinedModels-class.

DcLbmPrior(p = NaN)

DcLbm(alpha = 1, p = NaN)

Arguments

p

Exponential prior parameter (default to Nan, in this case p will be estimated from data as the average intensities of X)

alpha

Dirichlet prior parameter over the cluster proportions (default to 1)

Value

a DcLbmPrior-classa DcLbm-class object

See also

Examples

DcLbmPrior()
#> An object of class "DcLbmPrior"
#> Slot "p":
#> [1] NaN
#> 
DcLbmPrior(p = 0.7)
#> An object of class "DcLbmPrior"
#> Slot "p":
#> [1] 0.7
#> 
DcLbm()
#> An object of class "DcLbm"
#> Slot "alpha":
#> [1] 1
#> 
#> Slot "p":
#> [1] NaN
#> 
DcLbm(p = 0.7)
#> An object of class "DcLbm"
#> Slot "alpha":
#> [1] 1
#> 
#> Slot "p":
#> [1] 0.7
#>