Gaussian mixture model fit results class

An S4 class to represent a fit of a multivariate mixture of regression model, extend IclFit-class.

Slots

model

a GmmPrior-class object to store the model fitted

name

generative model name

icl

icl value of the fitted model

K

number of extracted clusters over row and columns

cl

a numeric vector with row and columns cluster indexes

obs_stats

a list with the following elements:

counts: numeric vector of size K with number of elements in each clusters

regs: list of size $K$ with statistics for each clusters

move_mat

binary matrix which store move constraints

train_hist

data.frame with training history information (details depends on the training procedure)

See also