Mixture of Multinomial Model Prior description class

An S4 class to represent a Mixture of Multinomial model. Such model can be used to cluster a data matrix \(X\) with the following generative model : $$ \pi \sim Dirichlet(\alpha)$$ $$ Z_i \sim \mathcal{M}(1,\pi)$$ $$ \theta_{k} \sim Dirichlet(\beta)$$ $$ X_{i.}|Z_{ik}=1 \sim \mathcal{M}(L_i,\theta_{k})$$ With \(L_i=\sum_d=1^DX_{id}\). These classes mainly store the prior parameters value (\(\alpha,\beta\)) of this generative model. The MoM-class must be used when fitting a simple Mixture of Multinomials whereas the MoMPrior-class must be sued when fitting a CombinedModels-class.

MoMPrior(beta = 1)

MoM(alpha = 1, beta = 1)

Arguments

beta

Dirichlet over vocabulary prior parameter (default to 1)

alpha

Dirichlet prior parameter over the cluster proportions (default to 1)

Value

a MoMPrior-class object a MoM-class object

See also

Examples

MoMPrior()
#> An object of class "MoMPrior"
#> Slot "beta":
#> [1] 1
#> 
MoMPrior(beta = 0.5)
#> An object of class "MoMPrior"
#> Slot "beta":
#> [1] 1
#> 
MoM()
#> An object of class "MoM"
#> Slot "alpha":
#> [1] 1
#> 
#> Slot "beta":
#> [1] 1
#> 
MoM(beta = 0.5)
#> An object of class "MoM"
#> Slot "alpha":
#> [1] 1
#> 
#> Slot "beta":
#> [1] 0.5
#>