Multinomial Stochastic Block Model hierarchical fit results class

An S4 class to represent a hierarchical fit of a Multinomial Stochastic Block Model, extend IclPath-class.

Slots

model

a MultSbm-class object to store the model fitted

name

generative model name

icl

icl value of the fitted model

K

number of extracted clusters over row and columns

cl

a numeric vector with row and columns cluster indexes

obs_stats

a list with the following elements:

counts: numeric vector of size K with number of elements in each clusters

x_counts: matrix of size KxKxM with the number of links between each pair of clusters

path

a list of size K-1 with each part of the path described by:

icl1: icl value reach with this solution for alpha=1

logalpha: log(alpha) value were this solution is better than its parent

K: number of clusters

cl: vector of cluster indexes

k,l: index of the cluster that were merged at this step

merge_mat: lower triangular matrix of delta icl values

obs_stats: a list with the elements:

counts: numeric vector of size K with number of elements in each clusters

x_counts: matrix of size KxKxM with the number of links between each pair of clusters

logalpha

value of log(alpha)

ggtree

data.frame with complete merge tree for easy plotting with ggplot2

tree

numeric vector with merge tree tree[i] contains the index of i father

train_hist

data.frame with training history information (details depends on the training procedure)

See also