Multinomial Stochastic Block Model Prior class

An S4 class to represent a Multinomial Stochastic Block Model. Such model can be used to cluster multilayer-graph vertex, and model a square adjacency cube \(X\) of size NxNxM with the following generative model : $$ \pi \sim Dirichlet(\alpha)$$ $$ Z_i \sim \mathcal{M}(1,\pi)$$ $$ \theta_{kl} \sim Dirichlet(\beta)$$ $$ X_{ij.}|Z_{ik}Z_{jl}=1 \sim \mathcal{M}(L_{ij},\theta_{kl})$$ With \(L_{ij}=\sum_{m=1}^MX_{ijm}\). These classes mainly store the prior parameters value \(\alpha,\beta\) of this generative model. The MultSbm-class must be used when fitting a simple MultSbm whereas the MultSbmPrior-class must be sued when fitting a MixedModels-class.

MultSbmPrior(beta = 1, type = "guess")

MultSbm(alpha = 1, beta = 1, type = "guess")

Arguments

beta
Dirichlet prior parameter over Multinomial links

type
define the type of networks (either "directed", "undirected" or "guess", default to "guess"), for undirected graphs the adjacency matrix is supposed to be symmetric.

alpha
Dirichlet prior parameter over the cluster proportions (default to 1)

Value

a MultSbmPrior-class object

a MultSbm-class object

Functions

MultSbm-class: MultSbm class constructor

MultSbmPrior: MultSbmPrior class constructor

MultSbm: MultSbm class constructor

Slots

beta

Dirichlet prior parameter over Multinomial links

type

define the type of networks (either "directed", "undirected" or "guess", default to "guess")

alpha

Dirichlet prior parameter over the cluster proportions (default to 1)

See also

Examples

MultSbmPrior()
#> An object of class "MultSbmPrior"
#> Slot "beta":
#> [1] 1
#> 
#> Slot "type":
#> [1] "guess"
#> 
MultSbmPrior(type = "undirected")
#> An object of class "MultSbmPrior"
#> Slot "beta":
#> [1] 1
#> 
#> Slot "type":
#> [1] "undirected"
#> 
MultSbm()
#> An object of class "MultSbm"
#> Slot "alpha":
#> [1] 1
#> 
#> Slot "beta":
#> [1] 1
#> 
#> Slot "type":
#> [1] "guess"
#> 
MultSbm(type = "undirected")
#> An object of class "MultSbm"
#> Slot "alpha":
#> [1] 1
#> 
#> Slot "beta":
#> [1] 1
#> 
#> Slot "type":
#> [1] "undirected"
#>