Degree corrected stochastic block model for bipartite graph hierarchical fit results class

An S4 class to represent a fit of a degree corrected stochastic block model for co_clustering, extend icl_path-class.

Slots

model

a co_dcsbm-class object to store the model fitted

name

generative model name

icl

icl value of the fitted model

K

number of extracted clusters over row and columns

Krow

number of extracted row clusters

Kcol

number of extracted column clusters

cl

a numeric vector with row and columns cluster indexes

obs_stats

a list with the following elements:

counts: numeric vector of size K with number of elements in each clusters

din: numeric vector of size K which store the sums of in-degrees for each clusters

dout: numeric vector of size K which store the sums of out-degrees for each clusters

x_counts: matrix of size K*K with the number of links between each pair of clusters

co_x_counts: matrix of size Krow*Kcol with the number of links between each pair of row and column cluster

clrow

a numeric vector with row cluster indexes

clcol

a numeric vector with column cluster indexes

Nrow

number of rows

Ncol

number of columns

path

a list of size K-1 with each part of the path described by:

icl1: icl value reach with this solution for alpha=1

logalpha: log(alpha) value were this solution is better than its parent

K: number of clusters

cl: vector of cluster indexes

k,l: index of the cluster that were merged at this step

merge_mat: lower triangular matrix of delta icl values

obs_stats: a list with the elements:

counts: numeric vector of size K with number of elements in each clusters

din: numeric vector of size K which store the sums of in-degrees for each clusters

dout: numeric vector of size K which store the sums of out-degrees for each clusters

x_counts: matrix of size K*K with the number of links between each pair of clusters

co_x_counts: matrix of size Krow*Kcol with the number of links between each pair of row and column cluster

logalpha

value of log(alpha)

ggtree

data.frame with complete merge tree for easy plotting with ggplot2

tree

numeric vector with merge tree tree[i] contains the index of i father

ggtreerow

data.frame with complete merge tree of row clusters for easy plotting with ggplot2

ggtreecol

data.frame with complete merge tree of column clusters for easy plotting with ggplot2

train_hist

data.frame with training history information (details depends on the training procedure)