Model based hierarchical clustering

Greed enable the clustering of networks and counts data such as document/term matrix with different model. Model selection and clustering is performed in combination by optimizing the Integrated Classification Likelihood (which is equivalent to minimizing the description length). Their following generative models are available :

Stochastic Block Model (directed)

Degree corrected Stochastic Block Model (directed)

Stochastic Block Model with Mutlinomial observations (experimental)

Mixture of Multinomial

Degree corrected Latent Block Model (directed)

Gaussian Mixture Model (experimental)

Mixture of Multivariate Regressions (experimental)

The optimization is performed thanks to a combination of greedy local search and a genetic algorithm. The main entry point is the greed function to perform the clustering.

greed(
  X,
  K = 20,
  model = find_model(X),
  alg = methods::new("hybrid"),
  verbose = FALSE
)

Arguments

X
data to cluster either a matrix,an array or a dgCMatrix-class

K
initial number of cluster

model
a generative model to fit sbm-class, dcsbm-class, co_dcsbm-class, mm-class or mvmreg-class

alg
an optimization algorithm of class hybrid-class (default), multistarts-class, seed-class or genetic-class

verbose
Boolean for verbose mode

Value

an icl_path-class object